backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
networkx==3.4.1
httpx==0.27.2
python-dotenv==1.0.1
openai==1.51.0
